requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
pyjwt==2.6.0
requests==2.28.1
python-dotenv==0.21.0
